dev-scheme/racket: first pass at 8.0 ebuild
Added {cs,bc,cgc} use flags. The default behavior is currently to
continue to install bc by default to match existing behavior. If
feedback says that cs should be made default then it will be in a
rev bump.

Issues related to pax-utils for the cs build have not been explored.

There is a known issue when installing cs with bc or cs with bc and
cgc where the call to run raco setup for cs does not propagate to the
instal. As a result startup times for cs in that case can be multiple
seconds. Currently working on a fix.