dev-db/pguri: fix gcc 14 compile errors
For >=dev-db/postgresql-16 and >=sys-devel/gcc-14 conventions
in the pg headers and in uri.c cause fatal errors, gcc 13 warns
in these situations as well, but is not fatal. There are two
known solutions, one is to fix uri.c as is done by the patch
in freebsd ports, the other is to pass -Wno-int-conversion
which is what we do for now since the issue appears to be in
both uri.c and in fmgr.h which would require patching postgres
itself. I suspect upstream will have to address this when they
start building with gcc 14.

https://github.com/petere/pguri/issues/16
https://github.com/freebsd/freebsd-ports/commit/68436150c0bbed44dec6393a310a9695cb869b89